rename some model members
