+++ src/main/java/org/icgc_argo/workflow/search/model/common/RunId.java
@@ -16,16 +16,15 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc_argo.workflow.search.rabbitmq;
+package org.icgc_argo.workflow.search.model.common;
 
-import lombok.Builder;
-import lombok.Getter;
-import org.icgc_argo.workflow.search.model.common.RunRequest;
+import lombok.*;
 
+@Data
 @Builder
-@Getter
-public class SenderDto {
-  String runId;
-  boolean cancelRequest;
-  RunRequest runRequest;
+@ToString
+@AllArgsConstructor
+@EqualsAndHashCode
+public class RunId {
+  @NonNull private String runId;
 }
